# api-docs/my-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "植物种植助手小程序后端API文档",
    "description": "植物种植助手小程序后端服务接口文档，基于Spring Boot 2.7.18实现。",
    "contact": {
      "name": "zszleon",
      "email": "2079973876@qq.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:18080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "TokenAuth": []
    }
  ],
...
  "paths": {
    "/api/reminders/update": {
      "post": {
        "tags": [
          "提醒管理"
        ],
        "summary": "更新提醒",
        "description": "更新提醒信息，需要Token认证",
        "operationId": "updateReminder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseReminder"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/reminders/page": {
      "post": {
        "tags": [
          "提醒管理"
        ],
        "summary": "获取提醒列表",
        "description": "分页获取当前用户的提醒列表，需要Token认证",
        "operationId": "getRemindersPage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderPageQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseReminder"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/reminders/isEnabled": {
      "post": {
        "tags": [
          "提醒管理"
        ],
        "summary": "切换提醒状态",
        "description": "启用或禁用提醒功能，需要Token认证",
        "operationId": "isEnabled",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IsEnabled"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseReminder"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/reminders/getReminderById": {
      "post": {
        "tags": [
          "提醒管理"
        ],
        "summary": "获取提醒详情",
        "description": "根据提醒ID获取提醒详细信息，需要Token认证",
        "operationId": "getReminderById",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Id"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseReminder"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/reminders/delete": {
      "post": {
        "tags": [
          "提醒管理"
        ],
        "summary": "删除提醒",
        "description": "删除指定提醒，需要Token认证",
        "operationId": "deleteReminder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Id"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/reminders/createReminder": {
      "post": {
        "tags": [
          "提醒管理"
        ],
        "summary": "创建提醒",
        "description": "创建新的养护提醒，需要Token认证",
        "operationId": "createReminder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseReminder"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponseObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "description": "统一API响应"
      },
      "ReminderRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "提醒ID，更新时必填",
            "format": "int64",
            "example": 123456789
          },
          "plantId": {
            "type": "integer",
            "description": "关联植物ID",
            "format": "int64",
            "example": 123456789
          },
          "type": {
            "type": "string",
            "description": "提醒类型",
            "example": "watering",
            "enum": [
              "watering",
              "fertilizing",
              "custom"
            ]
          },
          "customType": {
            "type": "string",
            "description": "自定义类型名称（当type为custom时）",
            "example": "修剪"
          },
          "time": {
            "type": "string",
            "description": "提醒时间（HH:mm格式）",
            "example": "09:00"
          },
          "frequency": {
            "type": "integer",
            "description": "频率（天数）",
            "format": "int32",
            "example": 7
          },
          "frequencyType": {
            "type": "string",
            "description": "频率类型",
            "example": "weekly",
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "custom"
            ]
          },
          "nextRemindTime": {
            "type": "string",
            "description": "下次提醒时间",
            "format": "date-time"
          },
          "isEnabled": {
            "type": "boolean",
            "description": "是否启用",
            "example": true
          }
        },
        "description": "提醒信息请求DTO"
      },
      "ApiResponseReminder": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/Reminder"
          }
        },
        "description": "统一API响应"
      },
      "Reminder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "plantId": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "customType": {
            "type": "string"
          },
          "time": {
            "type": "string"
          },
          "frequency": {
            "type": "integer",
            "format": "int32"
          },
          "frequencyType": {
            "type": "string"
          },
          "nextRemindTime": {
            "type": "string",
            "format": "date-time"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "响应数据"
      },
      "ReminderPageQueryRequest": {
        "type": "object",
        "properties": {
          "pageNum": {
            "minimum": 1,
            "type": "integer",
            "description": "页码",
            "format": "int32",
            "example": 1
          },
          "pageSize": {
            "minimum": 1,
            "type": "integer",
            "description": "每页条数",
            "format": "int32",
            "example": 10
          },
          "plantId": {
            "type": "integer",
            "description": "植物ID（可选）",
            "format": "int64",
            "example": 123456789
          }
        },
        "description": "提醒分页查询请求"
      },
      "ApiResponsePageResponseReminder": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseReminder"
          }
        },
        "description": "统一API响应"
      },
      "PageResponseReminder": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "description": "数据列表",
            "items": {
              "$ref": "#/components/schemas/Reminder"
            }
          },
          "total": {
            "type": "integer",
            "description": "总记录数",
            "format": "int64",
            "example": 100
          },
          "pageNum": {
            "type": "integer",
            "description": "当前页码",
            "format": "int32",
            "example": 1
          },
          "pageSize": {
            "type": "integer",
            "description": "每页条数",
            "format": "int32",
            "example": 10
          },
          "totalPages": {
            "type": "integer",
            "description": "总页数",
            "format": "int32",
            "example": 10
          }
        },
        "description": "分页响应"
      },
      "IsEnabled": {
        "required": [
          "id",
          "isEnabled"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "业务ID",
            "format": "int64",
            "example": 123456789
          },
          "isEnabled": {
            "type": "integer",
            "description": "是否启用（1-启用，0-停用）",
            "format": "int32",
            "example": 1,
            "enum": [
              0,
              1
            ]
          }
        },
        "description": "启用停用请求参数"
      },
      "Id": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "业务ID",
            "format": "int64",
            "example": 123456789
          }
        },
        "description": "ID请求参数"
      },
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "string",
            "description": "响应数据"
          }
        },
        "description": "统一API响应"
      },
...
    },
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "description": "用户认证Token，通过微信登录接口获取",
        "name": "X-Token",
        "in": "header"
      }
    }
  }
}